マウスイベント用モジュールのひな形実装
