+++ frontend/src/vue/pages/Islands/IslandsPage.vue
@@ -1,7 +1,18 @@
 <template>
   <div id="sightseeing-page" class="wrapper">
     <StatusTable :island="store.island" :status="store.status" :achievements="store.achievements" />
-    <IslandViewer />
+    <div id="island" ref="islandCanvas">
+      <TresCanvas v-bind="gl" id="island-canvas">
+        <IslandViewer>
+          <CameraControls />
+        </IslandViewer>
+      </TresCanvas>
+      <IslandHoverWindow
+        :showHoverWindow="store.showHoverWindow"
+        :hoverWindow="store.hoverWindow"
+        :hoverCellPoint="store.hoverCellPoint"
+        :terrains="store.terrains" />
+    </div>
     <div class="md:max-lg:px-3">
       <Bbs :island="store.island" />
       <LogViewer :title="store.island.name + '島の近況'" :parsed-logs="store.logs" />
@@ -13,7 +24,7 @@
 import StatusTable from '$vue/components/islands/common/StatusTable.vue'
 import LogViewer from '$vue/components/islands/common/LogViewer.vue'
 import IslandViewer from './IslandsViewer.vue'
-import { ref } from 'vue'
+import { onMounted, reactive, ref } from 'vue'
 import { Hakoniwa } from '$entity/Hakoniwa'
 import { Status } from '$entity/Status'
 import { Terrain } from '$entity/Terrain'
@@ -24,9 +35,11 @@ import { BbsMessage } from '$entity/Bbs'
 import Bbs from '$vue/components/islands/common/Bbs.vue'
 import { useIslandViewerStore } from '$store/IslandViewerStore.js'
 import { useBbsStore } from '$store/BbsStore.js'
-
-let hoverWindowTop = ref(170)
-let hoverWindowLeft = ref(0)
+import { TresCanvas } from '@tresjs/core'
+import IslandHoverWindow from '$vue/components/islands/Hover/IslandHoverWindow.vue'
+import { BasicShadowMap, NoToneMapping, SRGBColorSpace } from 'three'
+import CameraControls from '$vue/components/islands/Camera/CameraControls.vue'
+import { useCameraStore } from '$store/CameraControlsStore.js'
 
 const store = useIslandViewerStore()
 
@@ -50,9 +63,19 @@ interface Props {
 
 const props = defineProps<Props>()
 
+const gl = reactive({
+  clearColor: '#888888',
+  shadows: true,
+  alpha: true,
+  shadowMapType: BasicShadowMap,
+  outputColorSpace: SRGBColorSpace,
+  toneMapping: NoToneMapping
+})
+
 // Logsのparse
 const parser = new LogParser()
 const logs = parser.parse(props.island.logs, props.island.summary)
+const islandCanvas = ref<HTMLElement>()
 
 store.$patch((state) => {
   state.hakoniwa = props.hakoniwa
@@ -70,10 +93,58 @@ store.$patch((state) => {
 })
 
 useBbsStore().bbs = props.island.bbs
+
+onMounted(() => {
+  console.log(islandCanvas.value)
+  useCameraStore().canvas = islandCanvas.value
+  useCameraStore().init()
+})
 </script>
 
 <style lang="scss" scoped>
 #sightseeing-page {
   @apply mx-auto min-h-[1200px] max-w-[1000px] text-center;
+}
+
+#island {
+  margin: 0 auto;
+  @apply mb-4 w-full max-w-[512px] md:min-w-[512px];
+
+  #island-canvas {
+    @apply mb-4 max-h-[512px] min-h-[512px] w-full;
+  }
+
+  .row {
+    @apply m-0 bg-black p-0;
+    display: grid;
+
+    .cell {
+      @apply aspect-square w-full;
+    }
+
+    &:nth-child(odd) {
+      grid-template-columns: 1fr repeat(15, 2fr);
+    }
+
+    &:nth-child(even) {
+      grid-template-columns: repeat(15, 2fr) 1fr;
+    }
+
+    .left-padding {
+      @apply z-10 aspect-[1/2] w-full;
+      background-image: url('/img/hakoniwa/hakogif/land0.gif');
+      background-position: left;
+    }
+
+    .right-padding {
+      @apply relative z-10 aspect-[1/2] w-full;
+      background-image: url('/img/hakoniwa/hakogif/land0.gif');
+      background-position: right;
+
+      .right-padding-text {
+        @apply absolute left-1 z-10 overflow-hidden text-xs leading-none text-white max-xs:hidden md:text-sm;
+      }
+    }
+  }
 }
 </style>
